+++ db/migrate/20250222185950_add_boolean_is_machine_output_to_protocol.rb
@@ -1,0 +1,5 @@
+class AddBooleanIsMachineOutputToProtocol < ActiveRecord::Migration[7.2]
+  def change
+    add_column :protocols, :is_machine_output, :boolean
+  end
+end