+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Added
 
 - Append mode to multi update data attributes task [#4204]
+- <in progress Batch add/remove Protocols>
+- Protocol `is_machine_output` attribute to facilitate MachineObservation types
 
 [#4204]: https://github.com/SpeciesFileGroup/taxonworks/pull/4204